+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Version 0.3.2
+
+* Fix setting devid when calling `publish`
 # Version 0.3.1
 
 * Add `config` command to show your current Smaug configuration